# swagger-document.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/schedule/user/paginate": {
      "get": {
        "operationId": "UserController_paginate",
        "parameters": [
          {
            "name": "user_name",
            "required": false,
            "in": "query",
            "description": "Name of the user.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_email",
            "required": false,
            "in": "query",
            "description": "Email of the user.",
            "example": "levis@gmail.com",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone_number",
            "required": false,
            "in": "query",
            "description": "Phone number of the user.",
            "example": "997203320",
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "createdAt",
            "required": false,
            "in": "query",
            "description": "Date create of the user.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "updatedAt",
            "required": false,
            "in": "query",
            "description": "Date update of the user.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deletedAt",
            "required": false,
            "in": "query",
            "description": "Date delete of the user.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limite data of the paginate user.",
            "example": 100,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Current page of the paginate user.",
            "example": 1,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/schedule/user/{id}": {
      "get": {
        "operationId": "UserController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/schedule/user/findUserByEmail/{email}": {
      "get": {
        "operationId": "UserController_getByEmail",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/schedule/user": {
      "get": {
        "operationId": "UserController_list",
        "parameters": [
          {
            "name": "user_name",
            "required": false,
            "in": "query",
            "description": "Name of the user.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_email",
            "required": false,
            "in": "query",
            "description": "Email of the user.",
            "example": "levis@gmail.com",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone_number",
            "required": false,
            "in": "query",
            "description": "Phone number of the user.",
            "example": "997203320",
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "createdAt",
            "required": false,
            "in": "query",
            "description": "Date create of the user.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "updatedAt",
            "required": false,
            "in": "query",
            "description": "Date update of the user.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deletedAt",
            "required": false,
            "in": "query",
            "description": "Date delete of the user.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limite data of the paginate user.",
            "example": 100,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Current page of the paginate user.",
            "example": 1,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "UserController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/schedule/services/paginate": {
      "get": {
        "operationId": "ServicesController_paginate",
        "parameters": [
          {
            "name": "service_name",
            "required": false,
            "in": "query",
            "description": "Name of the service.",
            "example": "Degradê",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service_value",
            "required": false,
            "in": "query",
            "description": "Value of the service.",
            "example": 25,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "expected_time",
            "required": false,
            "in": "query",
            "description": "Time of the service.",
            "example": "25:00",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_active",
            "required": false,
            "in": "query",
            "description": "Is active of the service.",
            "example": true,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "user_id",
            "required": false,
            "in": "query",
            "description": "Affiliate user id (owner).",
            "example": 1,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "createdAt",
            "required": false,
            "in": "query",
            "description": "Date create of the service.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "updatedAt",
            "required": false,
            "in": "query",
            "description": "Date update of the service.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deletedAt",
            "required": false,
            "in": "query",
            "description": "Date delete of the service.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limite data of the paginate service.",
            "example": 100,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Current page of the paginate service.",
            "example": 1,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/schedule/services": {
      "get": {
        "operationId": "ServicesController_findAll",
        "parameters": [
          {
            "name": "service_name",
            "required": false,
            "in": "query",
            "description": "Name of the service.",
            "example": "Degradê",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service_value",
            "required": false,
            "in": "query",
            "description": "Value of the service.",
            "example": 25,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "expected_time",
            "required": false,
            "in": "query",
            "description": "Time of the service.",
            "example": "25:00",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_active",
            "required": false,
            "in": "query",
            "description": "Is active of the service.",
            "example": true,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "user_id",
            "required": false,
            "in": "query",
            "description": "Affiliate user id (owner).",
            "example": 1,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "createdAt",
            "required": false,
            "in": "query",
            "description": "Date create of the service.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "updatedAt",
            "required": false,
            "in": "query",
            "description": "Date update of the service.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deletedAt",
            "required": false,
            "in": "query",
            "description": "Date delete of the service.",
            "example": "2024-01-06",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limite data of the paginate service.",
            "example": 100,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Current page of the paginate service.",
            "example": 1,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Service"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ServicesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          }
        },
        "tags": [
          "services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/schedule/services/{id}": {
      "get": {
        "operationId": "ServicesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          }
        },
        "tags": [
          "services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "ServicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          }
        },
        "tags": [
          "services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ServicesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Schedule API example",
    "description": "The schedule API description",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "user_name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "user_email": {
            "type": "string"
          },
          "phone_number": {
            "type": "string",
            "nullable": true
          },
          "branchs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Branch"
            }
          },
          "id": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "user_name",
          "password",
          "user_email",
          "phone_number",
          "branchs",
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Client": {
        "type": "object",
        "properties": {
          "client_name": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "birth_date": {
            "type": "string"
          },
          "client_phone": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "branch_id": {
            "type": "number"
          },
          "branch": {
            "$ref": "#/components/schemas/Branch"
          },
          "id": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "client_name",
          "first_name",
          "birth_date",
          "client_phone",
          "is_active",
          "branch_id",
          "branch",
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Service": {
        "type": "object",
        "properties": {
          "service_name": {
            "type": "string"
          },
          "service_value": {
            "type": "number"
          },
          "expected_time": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "user_id": {
            "type": "number"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "branchs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Branch"
            }
          },
          "id": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "service_name",
          "service_value",
          "expected_time",
          "is_active",
          "user_id",
          "user",
          "branchs",
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Branch": {
        "type": "object",
        "properties": {
          "branch_name": {
            "type": "string"
          },
          "cnpj": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "cep": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "user_id": {
            "type": "number"
          },
          "district": {
            "type": "string"
          },
          "local_number": {
            "type": "string"
          },
          "branch_phone": {
            "type": "string"
          },
          "complements": {
            "type": "string"
          },
          "opening_hours": {
            "type": "string"
          },
          "closing_hours": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "clients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Client"
            }
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Service"
            }
          },
          "id": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "branch_name",
          "cnpj",
          "street",
          "cep",
          "city",
          "user_id",
          "district",
          "local_number",
          "branch_phone",
          "complements",
          "opening_hours",
          "closing_hours",
          "user",
          "clients",
          "services",
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "CreateUserDTO": {
        "type": "object",
        "properties": {
          "user_name": {
            "type": "string",
            "description": "Name of the user."
          },
          "password": {
            "type": "string",
            "description": "Password of the user.",
            "example": "909090"
          },
          "user_email": {
            "type": "string",
            "description": "Email of the user.",
            "example": "levis@gmail.com"
          },
          "phone_number": {
            "type": "string",
            "nullable": true,
            "description": "Phone number of the user.",
            "example": "997203320"
          }
        },
        "required": [
          "user_name",
          "password",
          "user_email"
        ]
      },
      "UpdateUserDTO": {
        "type": "object",
        "properties": {
          "user_name": {
            "type": "string",
            "description": "New name of the user."
          },
          "user_email": {
            "type": "string",
            "description": "New email of the user",
            "example": "levis@gmail.com"
          },
          "current_password": {
            "type": "string",
            "description": "Current password of the user",
            "example": "909090"
          },
          "password": {
            "type": "string",
            "description": "New password of the user",
            "example": "90909012"
          },
          "phone_number": {
            "type": "string",
            "nullable": true,
            "description": "New phone number of the user",
            "example": "997221050"
          }
        }
      },
...
      "CreateServiceDto": {
        "type": "object",
        "properties": {
          "service_name": {
            "type": "string",
            "description": "Name of the service.",
            "example": "Degradê"
          },
          "service_value": {
            "type": "number",
            "description": "Value of the service.",
            "example": 25
          },
          "expected_time": {
            "type": "string",
            "description": "Time of the service.",
            "example": "25:00"
          },
          "is_active": {
            "type": "boolean",
            "description": "Is active of the service.",
            "example": true
          },
          "user_id": {
            "type": "number",
            "description": "Affiliate user id (owner).",
            "example": 1
          }
        },
        "required": [
          "service_name",
          "service_value",
          "expected_time",
          "is_active",
          "user_id"
        ]
      },
      "UpdateServiceDto": {
        "type": "object",
        "properties": {
          "service_name": {
            "type": "string",
            "description": "Name of the service.",
            "example": "Degradê"
          },
          "service_value": {
            "type": "number",
            "description": "Value of the service.",
            "example": 25
          },
          "expected_time": {
            "type": "string",
            "description": "Time of the service.",
            "example": "25:00"
          },
          "is_active": {
            "type": "boolean",
            "description": "Is active of the service.",
            "example": true
          },
          "user_id": {
            "type": "number",
            "description": "Affiliate user id (owner).",
            "example": 1
          }
        }
      },
...
    }
  }
}